# HADOOP-19571. Improve PrometheusMetricsSink#normalizeName performance